+++ package.json
@@ -17,8 +17,8 @@
 		"db:generate": "drizzle-kit generate",
 		"db:migrate": "drizzle-kit push",
 		"format": "prettier --write '**/*.{js,ts,cjs,mjs,json}'",
-		"lint": "eslint .",
-		"lint:fix": "eslint . --fix",
+		"lint": "eslint . --config .github/linters/eslint.config.mjs",
+		"lint:fix": "eslint . --fix --config .github/linters/eslint.config.mjs",
 		"publish": "npx wrangler publish",
 		"start": "npx wrangler dev",
 		"up": "npx drizzle-kit up"
